% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{draw_rarefaction}
\alias{draw_rarefaction}
\title{Subsample species to draw a rarefaction curve}
\usage{
draw_rarefaction(population, pts = NA, replicates = 5, ecosystem = NA)
}
\arguments{
\item{population}{The population to draw from}

\item{pts}{A vector of depths to subsample to}

\item{replicates}{The number of times to repeat (for confidence intervals)}

\item{ecosystem}{The ecosystem name (useful for comparing ecosystems in plots)}
}
\value{
A data frame with each row corresponding to a sampling depth (number of rows = length of pts)
}
\description{
Subsample species to draw a rarefaction curve
}
