% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{get_error_rates}
\alias{get_error_rates}
\title{Get error rates for different approaches}
\usage{
get_error_rates(repeats, replicates, Ca, Cb, aa, ab, ba, bb, na1, na2, nb1, nb2)
}
\arguments{
\item{repeats}{}
}
\description{
Accepts a data structure and returns the estimated error rates for different approaches
}
