% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_point}
\alias{get_point}
\title{Points for plotting a rarefaction curved based on a sample}
\usage{
get_point(my_population, read, label, ecosystem_name, replicates = 1, pts)
}
\arguments{
\item{my_population}{The population distribution}

\item{read}{Sample depth}

\item{label}{label}

\item{ecosystem_name}{ecosystem_name}
}
\description{
Points for plotting a rarefaction curved based on a sample
}
