% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{make_community}
\alias{make_community}
\title{Generate a microbial community}
\usage{
make_community(CC, alpha_parameter, beta_parameter)
}
\arguments{
\item{CC}{Species richness of the community}

\item{alpha_parameter}{the alpha parameter for the beta distribution}

\item{beta_parameter}{the alpha parameter for the beta distribution}
}
\description{
Generate relative abundances of taxa from a beta distribution. Useful for illustrating and simulations.
}
