% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{rarefied_richness}
\alias{rarefied_richness}
\title{Rarefy a sample to a given level}
\usage{
rarefied_richness(my_sample, level)
}
\arguments{
\item{my_sample}{See sample_richness for details.}
}
\description{
Rarefy a sample to a given level
}
