% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefy.R
\name{rarefied_sample}
\alias{rarefied_sample}
\title{Rarefy a sample to a given level}
\usage{
rarefied_sample(smpl, level)
}
\arguments{
\item{smpl}{See sample_richness for details.}
}
\description{
Rarefy a sample to a given level
}
