% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_from_population}
\alias{sample_from_population}
\title{Sample species from a population}
\usage{
sample_from_population(population, sample_depth)
}
\arguments{
\item{population}{The names of the species}

\item{sample_depth}{The number of individuals to observe (the read depth)}
}
\value{
The species labels of the sampled species
}
\description{
Sample species from a population
}
