% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{sample_richness}
\alias{sample_richness}
\title{The observed richness of a sample}
\usage{
sample_richness(my_sample)
}
\arguments{
\item{my_sample}{A vector of the species labels. 
The vector should be of length n, where n is the number of indiviudals 
in the sample. The i-th element is the species label of individual i. 
If the relative abundance is given, a warning is thrown.}
}
\description{
The observed richness of a sample
}
