% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{get_error_rates}
\alias{get_error_rates}
\title{Estimate the probability of making the correct decision for different approaches}
\usage{
get_error_rates(repeats, replicates, Ca, Cb, aa, ab, ba, bb, na1, na2, nb1, nb2)
}
\arguments{
\item{repeats}{Number of times to run the simulation}
}
\description{
Accepts a data structure and returns the estimated error rates for different approaches
}
