% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{sample_from_population}
\alias{sample_from_population}
\title{Sample species from a population}
\usage{
sample_from_population(my_population_probabilities, sample_depth,
  my_population_names = NA)
}
\arguments{
\item{my_population_probabilities}{The relative abundances of the species}

\item{sample_depth}{The number of individuals to observe (the read depth)}

\item{my_population_names}{The names of the species}
}
\value{
The species labels of the sampled species
}
\description{
Sample species from a population
}
